# openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Machine API
  version: 0.1.1
  description: Simple coffee machine REST API.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is running.
  /readyz:
    get:
      summary: Readiness probe
      responses:
        '200':
          description: Machine can accept jobs.
        '503':
          description: Machine is busy.
  /status:
    get:
      summary: Machine status
      responses:
        '200':
          description: Machine is available.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '503':
          description: Machine is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /start-job:
    post:
      summary: Submit a brew job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartJobRequest'
      responses:
        '202':
          description: Job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid payload or unsupported product.
        '409':
          description: Job ID already exists.
        '503':
          description: Machine is busy.
  /retrieve-job:
    get:
      summary: Retrieve a brewed job
      parameters:
        - in: query
          name: jobID
          schema:
            type: string
          required: true
          description: Job identifier.
      responses:
        '200':
          description: Job is ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Missing jobID.
        '404':
          description: Job not found.
        '410':
          description: Job already retrieved.
        '503':
          description: Job not ready yet.
  /history:
    get:
      summary: Retrieve job history
      responses:
        '200':
          description: All known jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        '200':
          description: Coffee machine metrics.
          content:
            text/plain:
              schema:
                type: string
  /openapi.yaml:
    get:
      summary: OpenAPI specification
      responses:
        '200':
          description: Specification document.
          content:
            application/yaml:
              schema:
                type: string
components:
  schemas:
    StartJobRequest:
      type: object
      properties:
        jobId:
          type: string
          description: Optional client-provided job identifier.
        product:
          $ref: '#/components/schemas/Product'
      required:
        - product
    Job:
      type: object
      properties:
        jobId:
          type: string
        product:
          $ref: '#/components/schemas/Product'
        jobStarted:
          type: string
          format: date-time
        jobReady:
          type: string
          format: date-time
        jobRetrieved:
          type: string
          format: date-time
          nullable: true
      required:
        - jobId
        - product
        - jobStarted
        - jobReady
    Status:
      type: object
      properties:
        state:
          type: string
          enum:
            - available
            - brewing
            - blocked
        code:
          type: integer
      required:
        - state
        - code
    Product:
      type: string
      enum:
        - COFFEE
        - STRONG_COFFEE
        - CAPPUCCINO
        - COFFEE_WITH_MILK
        - ESPRESSO
        - ESPRESSO_CHOCOLATE
        - KAKAO
        - HOT_WATER